move files into package folder and use absolute imports
